street_english_app/requirements.txt
pandas
python-dotenv
eng_to_ipa
resend
openai